requirements.txt

# Stage 2 - Crawler
APScheduler==3.10.4
beautifulsoup4==4.12.2
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
gunicorn==21.2.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0

# Stage 4 - Spell Checker (migrated to Hunspell)
spylls==0.1.7

# Stage 5 - CTA Spell Checking
pyspellchecker==0.8.1
